% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_plotly.R
\name{column_plotly}
\alias{column_plotly}
\title{column_plotly Function}
\usage{
column_plotly(plot)
}
\arguments{
\item{plot}{A plotly from the server with covid tidy preprocessed data}
}
\value{
Returns a column with a plotly chart in the ui
}
\description{
This function generates a column in a fluidrow with a plotly inside
}
\examples{
column_plotly(plot = 'plotlytotalcases')

}
