% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidData.R
\docType{data}
\name{covidData}
\alias{covidData}
\title{Our World in Data's Covid Dataset}
\format{
a data.frame with 358480 observations and 4 variables
\itemize{
\item \strong{Country}: full country name
\item \strong{date}: the date of the day
\item \strong{variable}: the variable related to COVID-19, such as number of cases in the day, number of deaths and more
\item \strong{value}: The actual number that relates to the variable, the country and the date
}
}
\source{
\href{https://ourworldindata.org/coronavirus}{Our World in Data}
}
\usage{
covidData
}
\description{
A dataset containing the evolution of the variables related to Covid in 2020
Socioeconomic data is also available in the table but not used for the shiny APP
This dataset is the same of the preProcessedCovidData after melting operations - making it tidy
Only a few covid-related variables are explored in the shiny app
}
\keyword{datasets}
