% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch App Function}
\usage{
launch_app()
}
\value{
Covid Shiny App
}
\description{
This function launches the covid shiny app for the user
}
\examples{
\dontrun{
launch_app()
}


}
