% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases.R
\name{plot_cases}
\alias{plot_cases}
\title{plot_cases Function}
\usage{
plot_cases(covid_stat, label, title, country_compare)
}
\arguments{
\item{country_compare}{}
}
\value{
a plot using plotly and ggplot
}
\description{
plot_cases Function
}
