% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases.R
\name{plot_cases}
\alias{plot_cases}
\title{plot_cases Function}
\usage{
plot_cases(covid_stat, label, title, country_compare)
}
\arguments{
\item{covid_stat}{this is a description}

\item{label}{this is a description}

\item{title}{this is a description}

\item{country_compare}{this is a description}
}
\value{
a plot using plotly and ggplot
}
\description{
This function gives the user a stacked barchart plotly after
performing filtering in the dataset CovidData
}
\examples{
plot_cases(covid_stat = 'total_deaths_per_million', label = 'Date: \%s\nCountry: \%s\nDeaths: \%s', title = 'Accumulated Deaths per Million', country_compare = input$country_compare)

}
